Renamed the package recurrent_monthly_income_factory to just recurrent_income as the first set of renamings to a more reasonable name.
