+++ backend/src/main/java/entity/recurrent_income/CommonRecurrentMonthlyIncomeFactoryFactory.java
@@ -1,4 +1,4 @@
-package entity.recurrent_monthly_income_factory;
+package entity.recurrent_income;
 
 public class CommonRecurrentMonthlyIncomeFactoryFactory implements RecurrentMonthlyIncomeFactoryFactory {
     @Override
