+++ backend/src/main/java/entity/recurrent_income/RecurrentMonthlyIncomeFactoryFactory.java
@@ -1,4 +1,4 @@
-package entity.recurrent_monthly_income_factory;
+package entity.recurrent_income;
 
 public interface RecurrentMonthlyIncomeFactoryFactory {
